Refractored CLI and added tests for some parts
