+++ puzzle/cli/__init__.py
@@ -2,5 +2,9 @@ from .utils import (version, family_file, family_type, verbose, root,
  mode, variant_type)
 
 from .base import base
-from .store import init
+from .init import init
+from .delete import delete
+from .load import load
+from .individuals import individuals
+from .cases import cases
 from .view import view
